docker_submission/requirements.txt
torchio==0.19.6
nibabel==5.1.0
monai[nibabel]==1.3.0
monai-generative==0.2.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.11.1
pillow==9.4.0
seaborn==0.12.2
matplotlib==3.7.2
lpips
tensorboard
wandb
numba
pytorch-lightning
